refactor: reset the middlewares folder's structure
