+++ services/api-service/app/middlewares/validators/lucky-number.js
@@ -1,6 +1,6 @@
 import { createValidator } from '@utils/validator';
 
-export const createLuckyNumberValidator = createValidator({
+export const createLuckyNumberValidatorMiddleware = createValidator({
     key: [{ type: 'string', required: true, message: '活动标识是必需的' }],
     name: [{ type: 'string', required: true, message: '活动名称是必需的' }],
     description: [
@@ -9,22 +9,28 @@ export const createLuckyNumberValidator = createValidator({
     numbers: [{ type: 'array', required: true, message: '号码池数据是必需的' }],
 });
 
-export const queryLuckyNumberValidator = createValidator(
+export const queryLuckyNumberValidatorMiddleware = createValidator(
     {
         key: [{ type: 'string', required: true, message: '活动标识是必需的' }],
     },
     'params',
 );
 
-export const deleteLuckyNumberValidator = createValidator(
+export const deleteLuckyNumberValidatorMiddleware = createValidator(
     {
         key: [{ type: 'string', required: true, message: '活动标识是必需的' }],
     },
     'params',
 );
 
-export const cancelParticipatedLuckyNumberValidator = createValidator({
-    key: [{ type: 'string', required: true, message: '活动标识是必需的' }],
-    drawn_number: [{ type: 'number', required: true, message: '号码是必需的' }],
-    user_name: [{ type: 'string', required: true, message: '用户名是必需的' }],
-});
+export const cancelParticipatedLuckyNumberValidatorMiddleware = createValidator(
+    {
+        key: [{ type: 'string', required: true, message: '活动标识是必需的' }],
+        drawn_number: [
+            { type: 'number', required: true, message: '号码是必需的' },
+        ],
+        user_name: [
+            { type: 'string', required: true, message: '用户名是必需的' },
+        ],
+    },
+);
